requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
tensorflow==2.13.0
facenet-pytorch==2.5.3
albumentations==1.3.1
efficientnet-pytorch==0.7.1
numpy==1.24.3
pandas==2.1.0
matplotlib==3.8.0
opencv-python-headless==4.8.1.78
tqdm==4.66.1
Pillow==10.0.1
seaborn==0.13.0
scikit-learn==1.3.2
mtcnn==0.1.1
